Add DWARF for interface types
For Rust, the most natural way to represent a trait is using
DW_TAG_interface_type.  This patch adds support for interface types to
LLVM.

Note that this only adds interfaces to composite types -- but this is
insufficient, because Rust can have interfaces on subroutine, pointer,
and scalar types as well.